Codes/UpperMachine/requirements.txt
pyqt5
pyyaml
erniebot
pyqt5
matplotlib
pyserial # for desktop pet only
